Implement emulation of power-up handshake.
When powered up the radio or the panel will transmit the sequence
"0xFE 0xF0 0xFD" repeatedly until the other end responds with the
sequence "0xFE 0xF0 0xFD 0xFE 0xF1 0xFD".

The response is expected in less than a second, otherwise the unit will
power down again. Therefore, it is necessary to emulate this handshake
locally by sending the response right after the packet is received.