Improved power-on handshake emulation.
I noticed that if the panel is powered on while the radio is already on
and transmitting periodic updates, the panel will transmit the
PKT_TYPE_INIT2 sequenc and expect the same in response. If we don't
respond the panel will keep sending the packet forever generating a lot
of noise on the wire.