Add power button and control panel power management.
The power button is connected to GPIO7 using a pull-up resistor. The
button shorts the line to ground, thus the gpio is active low. The gpio
is set to trigger an interrupt on both rising and falling edges, which
allows reading it using the already exisiting select() call.

The power to the control panel is turned on and off using GPIO20.